rewrote conflate and fountain.properties to ts
