+++ server/common/build.info.ts
@@ -2,9 +2,9 @@
   // this file is automatically generated by git.version.js script
   const buildInfo = {
     version: '',
-    revision: '919b617',
+    revision: 'fb6f12c',
     branch: 'rewrite-services-to-typescript',
-    commit_time: '2021-03-19 11:02:34 +0100',
-    build_time: 'Mon Mar 22 2021 12:53:24 GMT+0100 (Central European Standard Time)'
+    commit_time: '2021-03-22 13:03:38 +0100',
+    build_time: 'Mon Mar 22 2021 18:06:47 GMT+0100 (Central European Standard Time)'
   };
   export default buildInfo;
