+++ server/common/constants.ts
@@ -14,3 +14,13 @@ export const PROP_STATUS_NOT_DEFINED = 'PROP_STATUS_NOT_DEFINED';
 export const PROP_STATUS_NOT_AVAILABLE = 'PROP_STATUS_NOT_AVAILABLE';
 export const MAX_IMG_SHOWN_IN_GALLERY = 50;
 export const LAZY_ARTIST_NAME_LOADING_i41db = true;
+
+export type PropStatus = 
+    'PROP_STATUS_OK' | 
+    'PROP_STATUS_INFO' | 
+    'PROP_STATUS_WARNING' | 
+    'PROP_STATUS_ERROR' |
+    'PROP_STATUS_FOUNTAIN_NOT_EXIST' |
+    'PROP_STATUS_NOT_DEFINED' |
+    'PROP_STATUS_NOT_AVAILABLE'
+;
