+++ tsconfig.json
@@ -2,7 +2,8 @@
   "compileOnSave": false,
   "compilerOptions": {
     "module": "commonjs",
-    "target": "es2020",
+     // TODO update once webpack bug regarding ?? is fixed
+     "target": "es2019",
     "lib": ["es2020"],
     "moduleResolution": "node",
     "outDir": "build",
